pyproject.toml
[tool.poetry]
name = "ashoka-buddhist-foundation-pro"
version = "0.1.0"
description = "High-performance website for The Ashoka Buddhist Foundation"
authors = ["Your Name <you@example.com>"] # Change this
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10" # Or newer
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = {extras = ["email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"} # SQLAlchemy 2.x for async
asyncpg = "^0.29.0" # PostgreSQL async driver
alembic = "^1.13.1"
jinja2 = "^3.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
redis = {extras = ["hiredis"], version = "^5.0.4"} # For caching and Celery
celery = {extras = ["redis"], version = "^5.3.6"}
flower = "^2.0.1" # Celery monitoring tool
httpx = "^0.27.0" # For async external API calls (if needed)
python-multipart = "^0.0.9" # For form data
greenlet = "^3.0.3" # SQLAlchemy async dependency
# Logging & Monitoring
structlog = "^24.1.0" # Example for structured logging
# opentelemetry-instrumentation-fastapi = "..." # For tracing (optional, advanced)
# sentry-sdk = {extras = ["fastapi"], version = "..."} # Error tracking

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
pytest-cov = "^5.0.0"
mypy = "^1.10.0"
ruff = "^0.4.4" # Linter & Formatter
pre-commit = "^3.7.0"
black = "^24.4.2" # Formatter (if not using ruff format)
httpx = "^0.27.0" # For test client

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 99
select = ["E", "F", "W", "I", "UP", "C4", "B", "A", "RUF"] # Enable various Ruff checks
ignore = ["E501"] # Handled by line-length

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true # Can be stricter in mature projects
plugins = ["pydantic.mypy"]